✨ feat: schedules model 분리
