+++ apps/studies/models/__init__.py
@@ -1,2 +1,3 @@
 from .groups import StudyGroup
 from .reviews import Review
+from .schedules import GroupSchedule, ScheduleParticipant
