+++ apps/studies/models/groups.py
@@ -1,8 +1,6 @@
 from django.conf import settings
-from django.core.exceptions import ValidationError
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
-from django.db.models import CheckConstraint, Q
 
 from apps.core.models import UUIDBaseModel
 
@@ -87,53 +85,3 @@ class GroupMember(UUIDBaseModel):
 
     def __str__(self) -> str:
         return f"{self.user.id} in group {self.study_group.id}"
-
-
-class GroupSchedule(UUIDBaseModel):
-    study_group = models.ForeignKey(
-        "StudyGroup",
-        on_delete=models.CASCADE,
-        related_name="schedules",
-    )
-    title = models.CharField(max_length=255, null=False, default="")
-    objective = models.CharField(max_length=255, null=False, default="")
-    session_date = models.DateField(null=False)
-    start_time = models.TimeField(null=False)
-    end_time = models.TimeField(null=False)
-
-    class Meta:
-        db_table = "group_schedules"
-        app_label = "studies"
-        constraints = [
-            CheckConstraint(
-                check=Q(end_time__gt=models.F("start_time")),
-                name="valid_schedule_time_range",
-            ),
-        ]
-
-    def clean(self) -> None:
-        if self.start_time >= self.end_time:
-            raise ValidationError("스터디 종료 시간은 시작 시간보다 늦어야 합니다.")
-
-
-class ScheduleParticipant(models.Model):
-    pk = models.CompositePrimaryKey("schedule_id", "member_id")
-
-    schedule = models.ForeignKey(
-        "GroupSchedule",
-        on_delete=models.CASCADE,
-        related_name="participants",
-        null=False,
-        blank=False,
-    )
-    member = models.ForeignKey(
-        "GroupMember",
-        on_delete=models.CASCADE,
-        related_name="schedules",
-        null=False,
-        blank=False,
-    )
-
-    class Meta:
-        db_table = "schedule_participants"
-        app_label = "studies"
